back/requirements.txt
fastapi
requests
langgraph
langchain_ollama
IPython
python-multipart
fastapi[standard]
psycopg2-binary
# TODO : fix the versions for the stable release